+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # v0.8.4 - WIP
 * bug fix
   * Re-enable session expiration detection
+* improvements
+  * Enable server exception handling
 
 # v0.8.3 - 2013-03-22
 * support Rails 3.2.13